backend/requirements.txt
amqp==5.2.0
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.9
django-allauth==0.57.0
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozendict==2.4.0
html5lib==1.1
idna==3.6
kombu==5.3.4
lxml==5.0.1
multitasking==0.0.11
numpy==1.26.3
oauthlib==3.2.2
pandas==2.1.4
peewee==3.17.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.35
